delete config.go
